chore(repo): 清理多余空行，修正类型转换并统一代码格式
- 删除多处函数内多余空行，提升代码整洁度和可读性
- 修正 freespace_unix.go 中的 uint64 和 int64 类型乘法顺序，避免隐式转换错误
- 优化 upload.go 中 int64 类型变量赋值，避免不必要的类型转换
- 统一代码风格，移除多余空行，遵循项目代码规范

+++ upload.go
@@ -86,7 +86,6 @@ func pickPartSize(size int64, maxParts int) (int, error) {
 
 func uploadFile(ctx context.Context, src string, size, mtime int64,
 	o uploadOpts) (tg.InputFileClass, error) {
-
 	name := filepath.Base(src)
 	if o.name != "" {
 		name = o.name
@@ -169,7 +168,7 @@ func uploadFile(ctx context.Context, src string, size, mtime int64,
 			continue
 		}
 
-		n := int64(min(int64(partSize), size-int64(idx)*int64(partSize)))
+		n := min(int64(partSize), size-int64(idx)*int64(partSize))
 		if o.pacer != nil {
 			if err := o.pacer.Acquire(ctx, n); err != nil {
 				setErr(err)
@@ -241,7 +240,6 @@ func uploadFile(ctx context.Context, src string, size, mtime int64,
 // sendPart 发送单个分片，带重试。调用方已持有一个信号量名额，这里负责归还。
 func sendPart(ctx context.Context, o uploadOpts, idx int, chunk []byte,
 	totalParts int, fileID int64, isBig bool, onDone func(int64)) error {
-
 	for attempt := 0; attempt < o.retries; attempt++ {
 		if ctx.Err() != nil {
 			return ctx.Err()
